requirements.txt
paho-mqtt
winrt
pyserial
keras
Pillow
numpy
opencv-python
tensorflow